# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Interacción con Base de Datos FireBird",
        "description": "API para la gestión",
        "version": "1.0"
    },
    "paths": {
...
        "/api/personas": {
            "get": {
                "tags": [
                    "Persona"
                ],
                "summary": "Obtener todas las personas",
                "operationId": "8601618afc784c4736e52022f7f6f047",
                "responses": {
                    "200": {
                        "description": "Lista de personas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Persona"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Persona"
                ],
                "summary": "Crear una nueva persona",
                "operationId": "d016dca10e1a596471fe5aa5d4c99841",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada exitosamente"
                    }
                }
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Persona"
                ],
                "summary": "Obtener una persona por ID",
                "operationId": "9b70ed1d108dd682ec86b7c48f328884",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información de la persona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Persona"
                ],
                "summary": "Actualizar una persona",
                "operationId": "2a6e07dc32d26f218b13c0a2dc442fdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Persona"
                ],
                "summary": "Eliminar una persona",
                "operationId": "8564ed4ed7e2d07b09bda4aac6184710",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Persona eliminada"
                    }
                }
            }
        },
...
        "/api/personas/restore/{id}": {
            "put": {
                "tags": [
                    "Persona"
                ],
                "summary": "Restaurar una persona eliminada",
                "operationId": "70bbdfbb88734a4dc46d7381423aa956",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona restaurada"
                    }
                }
            }
        },
        "/api/proveedores": {
            "get": {
                "tags": [
                    "Proveedor"
                ],
                "summary": "Obtener todos los proveedores",
                "operationId": "5eb7d790ae5c32c59bd58744f5ed3da5",
                "responses": {
                    "200": {
                        "description": "Lista de proveedores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Proveedor"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Proveedor"
                ],
                "summary": "Crear un nuevo proveedor",
                "operationId": "5c1bcfb0dbccf86c376dc7a47fdff311",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Proveedor"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor creado exitosamente"
                    }
                }
            }
        },
        "/api/proveedores/{id}": {
            "get": {
                "tags": [
                    "Proveedor"
                ],
                "summary": "Obtener un proveedor por ID",
                "operationId": "1898d8b0458ba55e2c8240363838a2a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proveedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del proveedor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proveedor"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Proveedor"
                ],
                "summary": "Actualizar un proveedor",
                "operationId": "72364f0ee2e9c4041a2f38af705cea3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proveedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Proveedor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proveedor actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Proveedor"
                ],
                "summary": "Eliminar un proveedor",
                "operationId": "8b7ae059b2d7afad943975d96233b0b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proveedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Proveedor eliminado"
                    }
                }
            }
        },
...
        "/api/proveedores/restore/{id}": {
            "put": {
                "tags": [
                    "Proveedor"
                ],
                "summary": "Restaurar un proveedor eliminado",
                "operationId": "1e36c165c25fd4abfe41f44be0412c9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proveedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor restaurado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Persona": {
                "title": "Persona",
                "description": "Modelo de Persona",
                "required": [
                    "CEDULA",
                    "NOMBRES",
                    "APELLIDO",
                    "TELEFONO",
                    "CORREO"
                ],
                "properties": {
                    "CEDULA": {
                        "description": "Cédula de la persona",
                        "type": "string"
                    },
                    "NOMBRES": {
                        "description": "Nombres de la persona",
                        "type": "string"
                    },
                    "APELLIDO": {
                        "description": "Apellido de la persona",
                        "type": "string"
                    },
                    "TELEFONO": {
                        "description": "Teléfono de la persona",
                        "type": "string"
                    },
                    "CORREO": {
                        "description": "Correo de la persona",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
            "Proveedor": {
                "title": "Proveedor",
                "description": "Modelo de Proveedor",
                "required": [
                    "NOMBRE",
                    "TELEFONO",
                    "CORREO",
                    "DIRECCION"
                ],
                "properties": {
                    "NOMBRE": {
                        "description": "Nombre del proveedor",
                        "type": "string"
                    },
                    "TELEFONO": {
                        "description": "Teléfono del proveedor",
                        "type": "string"
                    },
                    "CORREO": {
                        "description": "Correo del proveedor",
                        "type": "string"
                    },
                    "DIRECCION": {
                        "description": "Dirección del proveedor",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}